requirements.txt
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
distro==1.5.0
Flask==1.1.2
idna==2.10
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.19.0
pandas==1.0.5
pylint==2.5.3
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
tabula-py==2.1.1
toml==0.10.1
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
XlsxWriter==1.2.9
